+++ HISTORY.rst
@@ -3,6 +3,12 @@
 History
 -------
 
+1.1.0 (2020-10-27)
+++++++++++++++++++
+
+* Set default value to ``'no-referrer-when-downgrade'`` instead of requiring it
+* Add Django 3.1 and Python 3.9 for testing
+
 1.0.1 (2019-08-23)
 ++++++++++++++++++